+++ src/wayland.ts
@@ -1,0 +1,4 @@
+import Logger from "./logging.ts";
+import Options from "./options.ts";
+
+export default function startX11(options: Options, logger: Logger) {}
